% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig.R
\name{gig}
\alias{gig}
\alias{rgig}
\alias{dgig}
\title{Generalised Inverse Gaussian}
\usage{
dgig(x, a, b, p, log_density)

rgig(n, lambda, a, b)
}
\arguments{
\item{x}{: vector of points}

\item{a}{: chi parameter}

\item{b}{: psi parameter}

\item{log_density}{: returns log density}

\item{n}{: number of observations}

\item{lambda, p}{: lambda parameter}
}
\value{
\itemize{
\item {\code{rgig}} {returns a vector of random numbers from GIG(a,b,p).}
\item {\code{dgig}} {returns returns density of a GIG(a,b,p) at point x.}
}
}
\description{
Probability distribution function, Quantile Random generation 
for the Generalised Inverse Gaussian with three parameters a,b,p.
}
\details{
The Generalised Inverse Gaussian distrubtion(GIG), which has the following pdf

\deqn{f(x) = x^{\lambda-1}exp{-\frac{\omega}{2}(x + \frac{1}{x})}}{f(x) = x^(\lambda-1) exp{-\omega/2 (x + 1/x)}}
}
\examples{
rgig(1,0.5, 1, 2)
dgig(1, 1, 2, 0.5, FALSE)

}
\references{
Devroye, L. Random variate generation for the generalized inverse Gaussian distribution. 
Stat Comput 24, 239–246 (2014).
}
