% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbldclass.R
\name{plot.qbld}
\alias{plot.qbld}
\title{Plot QBLD}
\usage{
\method{plot}{qbld}(
  x,
  trace = TRUE,
  density = TRUE,
  auto.layout = TRUE,
  ask = dev.interactive(),
  ...
)
}
\description{
Plots `qbld` class object.
}
\details{
plot.qbld has following options:
\itemize{
\item {\code{trace}} {Whether or not to plot trace}
\item {\code{density}}  {Whether or not to plot density}
\item {\code{auto.layout}} {Auto set layout or not}
}
}
