% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbldclass.R
\name{summary.qbld}
\alias{summary.qbld}
\alias{print.summary.qbld}
\title{QBLD Summary Class}
\usage{
\method{summary}{qbld}(
  object,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  epsilon = 0.05,
  ...
)

\method{print}{summary.qbld}(x, ...)
}
\arguments{
\item{object}{: `qbld` class object}

\item{quantiles}{: Vector of quantiles for summary of the covariates, defaulted to c(0.025, 0.25, 0.5, 0.75, 0.975)}

\item{epsilon}{: epsilon value for calculating target.psrf and significiance stars, 0.05 by default.}

\item{...}{: Other summary arguments}

\item{x}{: (for print.summary.qbld) `qbld.summary` class object}
}
\value{
summary.qbld produces following sets of summary statistics for each variable:
\itemize{
\item {\code{statistics}} {Contains the mean, sd, markov std error, ess and gelman-ruben diagnostic}
\item {\code{quantiles}}  {Contains quantile estimates for each variable}
\item {\code{nsim}}       {No. of simulations run}
\item {\code{burn}}       {Burn-in used or not}
\item {\code{which}}      {Block, or Unblock version of sampler}
\item {\code{p}}          {quantile for the AL distribution on the error term}
\item {\code{multiess}}   {multiess value for the sample}
\item {\code{multigelman}} {multivariate version of gelman-ruben}
}
}
\description{
Outputs a `summary.qbld` class object, and prints as described.
}
