% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbldclass.R
\name{plot.qbld}
\alias{plot.qbld}
\title{Plot QBLD}
\usage{
\method{plot}{qbld}(x,trace = TRUE,density = TRUE,auto.layout = TRUE,ask = dev.interactive(),...)
}
\arguments{
\item{x}{: `qbld' class object to plot.}

\item{trace}{: Whether or not to plot trace plots for covariates, TRUE by default}

\item{density}{: Whether or not to plot density for covariates, TRUE by default.}

\item{auto.layout}{: Auto set layout or not, TRUE as default. Plots according to the local settings if false.}

\item{ask}{: For Interactive plots}

\item{...}{: Other plot arguments}
}
\value{
Plots as specified.
}
\description{
Plots `qbld' class object.
}
\seealso{
\code{\link{summary.qbld}}, \code{\link{model.qbld}}
}
