\name{qbld-package}
\alias{qbld-package}
\alias{qbld}
\docType{package}
\title{
\packageTitle{qbld}
}
\description{
\packageDescription{qbld}
}

\author{
\packageAuthor{qbld}

Maintainer: \packageMaintainer{qbld}
}
\references{
Rahman, Mohammad & Vossmeyer, Angela. (2019). Estimation and Applications of Quantile Regression for Binary Longitudinal Data. Advances in Econometrics. 40.
}
\keyword{package}

